move dashboard specific code to dashboards package
