+++ libbeat/dashboards/index_pattern.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package kibana
+package dashboards
 
 import (
 	"bufio"
